refactor: Move update-related classes and add `UpdateProgress` event
- Moved `AppVersion` and `UpdateEvents` into a new `viewmodel` subpackage.
- Introduced the `UpdateProgress` event to the `UpdateEvents` sealed class.
